final scrutinizer for the foreseeable

+++ lib/class/rating.class.php
@@ -56,7 +56,7 @@ class Rating extends database_object
     {
         $types = array('song', 'album', 'artist', 'video', 'tvshow', 'tvshow_season', 'playlist', 'label', 'podcast', 'podcast_episode');
 
-        if ($object_type != null) {
+        if ($object_type !== null && $object_type !== '') {
             if (in_array($object_type, $types)) {
                 $sql = "DELETE FROM `rating` WHERE `object_type` = ? AND `object_id` = ?";
                 Dba::write($sql, array($object_type, $object_id));
